fix: minor fixes related to maxFrameBytes and new dependency injection

+++ Server/src/main/kotlin/com/oeims/connections/WebSocketManager.kt
@@ -70,7 +70,7 @@ class WebSocketManager<K: Any> {
     suspend fun receiveText(
         key: K,
         session: DefaultWebSocketServerSession,
-        maxFrameBytes: Int? = null,
+        maxFrameBytes: Long? = null,
         onTextFrame: suspend (String) -> Unit,
     ) {
         try {
